add fsdj apis

+++ pkg/file/file.go
@@ -92,20 +92,20 @@ func MustOpen(fileName, filePath string) (*os.File, error) {
 	return f, nil
 }
 
-//judement file modtime older than overDate
+// Judement file modtime older than overDate
 func IsOlderThanDate(t time.Time, overDate int64) bool {
 	return time.Now().Sub(t) > time.Duration(overDate)*24*time.Hour
 }
 
-//find files older than overDate
-func FindFilesOlderThanDate(dir string, overDate int64) (files []os.FileInfo, err error) {
+// Find files older than overDate
+func FindFilesOlderThanDate(dir string, overDate int) (files []os.FileInfo, err error) {
 	tmpfiles, err := ioutil.ReadDir(dir)
 	if err != nil {
 		return nil, err
 	}
 	for _, file := range tmpfiles {
 		if file.Mode().IsRegular() {
-			if IsOlderThanDate(file.ModTime(), overDate) {
+			if IsOlderThanDate(file.ModTime(), int64(overDate)) {
 				files = append(files, file)
 			}
 		}
